chore(sdk): move utils into shared, add documentation for submodules
