+++ libs/wingsdk/src/cloud/function.ts
@@ -3,8 +3,8 @@ import { join } from "path";
 import { Construct } from "constructs";
 import { fqnForType } from "../constants";
 import { App } from "../core/app";
+import { CaseConventions, ResourceNames } from "../core/resource-names";
 import { Duration, IInflightHost, IResource, Resource } from "../std";
-import { CaseConventions, ResourceNames } from "../utils/resource-names";
 
 /**
  * Global identifier for `Function`.
