+++ libs/wingsdk/src/core/resource-names.ts
