+++ libs/wingsdk/src/shared/bundling.ts
