+++ libs/wingsdk/src/shared/table-utils.ts
