+++ libs/wingsdk/src/target-sim/redis.inflight.ts
@@ -2,11 +2,11 @@ import IoRedis from "ioredis";
 import { v4 as uuidv4 } from "uuid";
 import { RedisAttributes, RedisSchema } from "./schema-resources";
 import { RedisClientBase } from "../redis";
+import { runCommand } from "../shared/misc";
 import {
   ISimulatorContext,
   ISimulatorResourceInstance,
 } from "../testing/simulator";
-import { runCommand } from "../utils/misc";
 
 export class Redis
   extends RedisClientBase
