+++ libs/wingsdk/src/target-sim/secret.ts
@@ -4,9 +4,9 @@ import { SECRET_TYPE, SecretSchema } from "./schema-resources";
 import { bindSimulatorResource, makeSimulatorJsClient } from "./util";
 import * as cloud from "../cloud";
 import * as core from "../core";
+import { ResourceNames } from "../core/resource-names";
 import { IInflightHost } from "../std";
 import { BaseResourceSchema } from "../testing/simulator";
-import { ResourceNames } from "../utils/resource-names";
 
 /**
  * Simulator implementation of `cloud.Secret`
