+++ libs/wingsdk/src/target-tf-aws/bucket.ts
@@ -17,13 +17,13 @@ import { S3Object } from "../.gen/providers/aws/s3-object";
 import * as cloud from "../cloud";
 import { BucketEventType, Topic } from "../cloud";
 import * as core from "../core";
-import { calculateBucketPermissions } from "../shared-aws/permissions";
-import { IInflightHost } from "../std";
 import {
   CaseConventions,
   NameOptions,
   ResourceNames,
-} from "../utils/resource-names";
+} from "../core/resource-names";
+import { calculateBucketPermissions } from "../shared-aws/permissions";
+import { IInflightHost } from "../std";
 
 const EVENTS = {
   [BucketEventType.DELETE]: ["s3:ObjectRemoved:*"],
