+++ libs/wingsdk/src/target-tf-aws/counter.ts
@@ -3,10 +3,10 @@ import { Function } from "./function";
 import { DynamodbTable } from "../.gen/providers/aws/dynamodb-table";
 import * as cloud from "../cloud";
 import * as core from "../core";
+import { NameOptions, ResourceNames } from "../core/resource-names";
 import { COUNTER_HASH_KEY } from "../shared-aws/commons";
 import { calculateCounterPermissions } from "../shared-aws/permissions";
 import { IInflightHost } from "../std";
-import { NameOptions, ResourceNames } from "../utils/resource-names";
 
 /**
  * Counter (Table) names must be between 3 and 255 characters.
